Rename file to `LocalSourceUpload`
